fix bug with int, replace with uint64_t

Former-commit-id: 2ccd906c37e2d0f30f9f130e8fd5f6198d59d364

+++ src/kmer-index/bitarray.h
@@ -1,13 +1,14 @@
 #include<stdio.h>
 #include<stdlib.h>
+#include <stdint.h>
 
 typedef struct {
   char* values;
-  int size;
-  int capacity;
+  uint64_t size;
+  uint64_t capacity;
 } bitarray_t;
 
-bitarray_t* create_bitarray(int n);
+bitarray_t* create_bitarray(uint64_t n);
 
 void add_to_bitarray(bitarray_t* array, int value);
